Remote updating from the Web now works
This is the rough draft of a working remote update. While fw_update is working
the web page will at like its loading until fw_update is complete, then a log will print out
and the page will be come usable. Working on a better way to let user know what is happening during
this period. Rebooting will soon be added to the web as well.