WebLCM: Active profile is now clearly displayed above the profile list, sdc_cli secuirty mismatch
Reworked how the profile list is displayed. This way is simpler and more effective.
Now the active profile is known before the list is generated instead of during.

Security type is different from what set weptype accepts and what scan returns.
Catch and reset this to a valid input.

Bugs: 5472, 5492